[refactor] #61 폴더링 변경
